fix: EditMarkerWindow moved to a more fittingly named file
