+++ src/render/mod.rs
@@ -4,7 +4,7 @@ pub mod info_tab;
 #[cfg(feature = "markers")]
 pub mod marker_tab;
 #[cfg(feature = "markers-edit")]
-pub mod marker_window;
+pub mod edit_marker_window;
 pub mod primary_window;
 pub mod state;
 pub mod timer_tab;
@@ -12,6 +12,10 @@ pub mod timer_window;
 
 #[cfg(feature = "markers")]
 pub use marker_tab::MarkerTabState;
+
+#[cfg(feature = "markers-edit")]
+pub use edit_marker_window::EditMarkerWindowState;
+
 #[allow(unused_imports)]
 pub use {
     config_tab::ConfigTabState,
